Refactor stand rerenderer.
