+++ src/Utility/Extensions.cs
@@ -1,32 +1,10 @@
 using System.Collections.Generic;
-using DoffAndDonAgain.Common;
-using DoffAndDonAgain.Common.Network;
 using Vintagestory.API.Common;
 using Vintagestory.API.Common.Entities;
-using Vintagestory.API.MathTools;
-using Vintagestory.API.Server;
 using Vintagestory.GameContent;
 
 namespace DoffAndDonAgain.Utility {
   public static class GameExtensions {
-
-    #region EntityArmorStand
-    public static void UpdateRender(this EntityArmorStand armorStand) {
-      if (armorStand.World.Side == EnumAppSide.Server) {
-        armorStand.WatchedAttributes.MarkAllDirty();
-        armorStand.World.RegisterCallback(armorStand.RenderUpdateCallback, armorStand.Pos.AsBlockPos, 500);
-      }
-      else if (armorStand.IsRendered) {
-        armorStand.OnEntityLoaded();
-      }
-    }
-
-    private static void RenderUpdateCallback(this EntityArmorStand armorStand, IWorldAccessor world, BlockPos pos, float dt) {
-      // TODO: having to get my channel here probably means this should be inside the rerender handler instead?
-      (world?.Api as ICoreServerAPI)?.Network.GetChannel(Constants.MOD_ID).BroadcastPacket(new ArmorStandInventoryUpdatedPacket(armorStand.EntityId));
-    }
-
-    #endregion
 
     #region EntityAgent
